refactor: CarRepository 객체 네이밍 변경(Participants)
