+++ src/test/java/repository/ParticipantsTest.java
@@ -1,18 +1,19 @@
 package repository;
 
 import domain.Car;
+import domain.Participants;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class CarRepositoryTest {
+class ParticipantsTest {
 
-    private CarRepository carRepository;
+    private Participants participants;
 
     @BeforeEach
     void before() {
-        carRepository = new CarRepository();
+        participants = new Participants();
     }
 
     @Test
@@ -22,11 +23,11 @@ class CarRepositoryTest {
         final String name2 = "asdf";
         List<Car> cars = List.of(new Car(name1), new Car(name1), new Car(name2));
         //when
-        carRepository.add(name1);
-        carRepository.add(name1);
-        carRepository.add(name2);
+        participants.add(name1);
+        participants.add(name1);
+        participants.add(name2);
         //then
-        List<Car> results = carRepository.findAll();
+        List<Car> results = participants.findAll();
         Assertions.assertThat(results.size()).isEqualTo(3);
         Assertions.assertThat(results.get(1).getName()).isEqualTo(cars.get(1).getName());
         Assertions.assertThat(results.get(0).getName()).isEqualTo(cars.get(0).getName());
